+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: asuc <asuc@student.42angouleme.fr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 18:49:06 by asuc              #+#    #+#             */
-/*   Updated: 2023/11/09 18:06:45 by asuc             ###   ########.fr       */
+/*   Updated: 2023/11/10 18:21:39 by asuc             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@
 # include "../libft/libft.h"
 
 int		ft_printf(const char *format, ...);
-char	*ft_itoa_base(unsigned long int n, int base);
-int		ft_num_len_base(unsigned long int n, int base);
-int		print_hex(unsigned long long nb);
+char	*ft_itoa_base(unsigned long n, int base);
+int		ft_num_len_base(unsigned long n, int base);
+int		print_hex(unsigned long nb);
 int		print_hex_other(unsigned int nb, int mode);
 int		print_format(const char *format, int index, va_list arg);
 char	*ft_itoa_base_hex(unsigned int n, int base, int mode);
